+++ src/app/GeneratorTeamsApp.ts
@@ -170,6 +170,7 @@ export class GeneratorTeamsApp extends Generator {
                 process.exit(0)
             }
             if (!this.options.existingManifest) {
+
                 answers.host = answers.host.endsWith('/') ? answers.host.substr(0, answers.host.length - 1) : answers.host;
                 this.options.title = answers.name;
                 this.options.description = this.description;
@@ -180,10 +181,10 @@ export class GeneratorTeamsApp extends Generator {
                 this.options.developer = answers.developer;
                 this.options.host = answers.host;
                 var tmp: string = this.options.host.substring(this.options.host.indexOf('://') + 3)
+                this.options.hostname = this.options.host.substring(this.options.host.indexOf('://') +3);
+
                 var arr: string[] = tmp.split('.');
                 this.options.namespace = lodash.reverse(arr).join('.');
-                this.options.tou = answers.host + '/tou.html';
-                this.options.privacy = answers.host + '/privacy.html';
                 this.options.id = Guid.raw();
                 if (this.options.host.indexOf('azurewebsites.net') >= 0) {
                     this.options.websitePrefix = this.options.host.substring(this.options.host.indexOf('://') + 3, this.options.host.indexOf('.'));