+++ src/custombot/templates/src/app/{customBotName}.ts
@@ -23,10 +23,12 @@ export class <%= customBotName %> implements IOutgoingWebhook {
      */
     public requestHandler(req: express.Request, res: express.Response, next: express.NextFunction) {
         // parse the incoming message
-        const incoming = <builder.IMessage>req.body
+        const incoming = <builder.Activity>req.body
 
         // create the response, any Teams compatible responses can be used
-        const message = new builder.Message();
+        const message: Partial<builder.Activity> = {
+            type: builder.ActivityTypes.Message
+        };
     
         const securityToken = process.env.SECURITY_TOKEN;
         if (securityToken && securityToken.length > 0) {
@@ -40,17 +42,17 @@ export class <%= customBotName %> implements IOutgoingWebhook {
 
             if (msgHash === auth) {
                 // Message was ok and verified
-                message.text(`Echo ${incoming.text}`);
+                message.text = `Echo ${incoming.text}`;
             } else {
                 // Message could not be verified
-                message.text(`Error: message sender cannot be verified`);
+                message.text = `Error: message sender cannot be verified`;
             }
         } else {
             // There is no configured security token
-            message.text(`Error: outgoing webhook is not configured with a security token`);
+            message.text = `Error: outgoing webhook is not configured with a security token`;
         }
 
         // send the message
-        res.send(message.toMessage());
+        res.send(JSON.stringify(message));
     }
 }