+++ src/messageExtension/MessageExtensionGenerator.ts
@@ -248,12 +248,12 @@ export class MessageExtensionGenerator extends Generator {
                 });
 
                 Yotilities.addAdditionalDeps([
-                    ["msteams-ui-components-react", "^0.7.3"],
-                    ["react", "^16.1.0"],
-                    ["@types/react", "16.4.7"],
-                    ["react-dom", "^16.2.0"],
+                    ["msteams-ui-components-react", "^0.8.1"],
+                    ["react", "^16.8.4"],
+                    ["@types/react", "16.8.8"],
+                    ["react-dom", "^16.8.4"],
                     ["file-loader", "1.1.11"],
-                    ["typestyle", "1.5.1"]
+                    ["typestyle", "2.0.1"]
                 ], this.fs);
 
                 Yotilities.insertTsExportDeclaration(
@@ -272,7 +272,7 @@ export class MessageExtensionGenerator extends Generator {
 
                 // Dynamically insert the reference and hook it up to the Bot
                 const project = new Project();
-                const file = project.createSourceFile(`src/app/${this.options.botName}.ts`, this.fs.read(`src/app/${this.options.botName}.ts`), {
+                const file = project.createSourceFile(`src/app/${this.options.botName}/bot.ts`, this.fs.read(`src/app/${this.options.botName}/bot.ts`), {
                     overwrite: true
                 });
                 const classes = file.getClasses()
@@ -285,7 +285,7 @@ export class MessageExtensionGenerator extends Generator {
                 const pos = lastImport !== undefined ? lastImport.getChildIndex() : 0;
                 const importDecl = file.insertImportDeclaration(pos, {
                     defaultImport: this.options.messageExtensionName,
-                    moduleSpecifier: `./${this.options.messageExtensionName}`,
+                    moduleSpecifier: `../${this.options.messageExtensionName}`,
                 });
                 let hostimports = imports.filter(i => {
                     return i.getModuleSpecifier().getLiteralText() == 'express-msteams-host'
@@ -310,8 +310,9 @@ export class MessageExtensionGenerator extends Generator {
                         scope: Scope.Private,
                         name: `_${this.options.messageExtensionName}`,
                         type: this.options.messageExtensionName,
-                        docs: [`Local property for ${this.options.messageExtensionName}`],
+                        docs: [`Local property for ${this.options.messageExtensionName}`]
                     });
+                    
                     // add the decorator
                     prop.addDecorator({
                         name: 'MessageExtensionDeclaration',
@@ -323,22 +324,9 @@ export class MessageExtensionGenerator extends Generator {
                     const constructors = cl.getConstructors();
                     if (constructors.length > 0) {
                         const c = constructors[0];
-
-                        let statements = c.getStatements().filter(s => {
-                            return s.getText() == 'super.registerMessageExtensions();';
-                        });
-
-                        if (statements.length == 1) {
-                            // insert the variable
-                            c.insertStatements(statements[0].getChildIndex(), `// Message extension ${this.options.messageExtensionName}
-                            this._${this.options.messageExtensionName} = new ${this.options.messageExtensionName}(this.universalBot);`);
-                        } else {
-                            // insert variable and registrations
-                            c.addStatements(`// Message extension ${this.options.messageExtensionName}
-                            this._${this.options.messageExtensionName} = new ${this.options.messageExtensionName}(this.universalBot);
-                            // Register all messge extensions
-                            super.registerMessageExtensions();`)
-                        }
+                        c.insertStatements(0, `// Message extension ${this.options.messageExtensionName}
+                        this._${this.options.messageExtensionName} = new ${this.options.messageExtensionName}();
+                        `);
 
                     } else {
                         // TODO: log
@@ -347,7 +335,7 @@ export class MessageExtensionGenerator extends Generator {
                     // TODO: log
                 }
                 file.formatText();
-                this.fs.write(`src/app/${this.options.botName}.ts`, file.getFullText());
+                this.fs.write(`src/app/${this.options.botName}/bot.ts`, file.getFullText());
             }
         }
     }