+++ src/tab/TabGenerator.ts
@@ -50,12 +50,12 @@ export class TabGenerator extends Generator {
     public writing() {
         if (this.options.tab) {
             let templateFiles = [
-                "src/app/scripts/{tabName}Config.tsx",
-                "src/app/scripts/{tabName}.tsx",
-                "src/app/scripts/{tabName}Remove.tsx",
-                "src/app/web/{tabName}.html",
-                "src/app/web/{tabName}Remove.html",
-                "src/app/web/{tabName}Config.html",
+                "src/app/scripts/{tabName}/{tabName}Config.tsx",
+                "src/app/scripts/{tabName}/{tabName}.tsx",
+                "src/app/scripts/{tabName}/{tabName}Remove.tsx",
+                "src/app/web/{tabName}/{tabName}.html",
+                "src/app/web/{tabName}/{tabName}Remove.html",
+                "src/app/web/{tabName}/{tabName}Config.html",
             ];
 
             this.sourceRoot()
@@ -73,41 +73,37 @@ export class TabGenerator extends Generator {
             let manifestPath = "src/manifest/manifest.json";
             var manifest: any = this.fs.readJSON(manifestPath);
             (<any[]>manifest.configurableTabs).push({
-                configurationUrl: `${this.options.host}/${this.options.tabName}Config.html`,
+                configurationUrl: `https://{{HOSTNAME}}/${this.options.tabName}/${this.options.tabName}Config.html`,
                 canUpdateConfiguration: true,
                 scopes: ["team"]
             });
-            var tmp: string = this.options.host.substring(this.options.host.indexOf('://') + 3)
-            var arr: string[] = tmp.split('.');
-            ;
-            (<string[]>manifest.validDomains).push(this.options.host.split("https://")[1]);
             this.fs.writeJSON(manifestPath, manifest);
 
             Yotilities.addAdditionalDeps([
-                ["msteams-ui-components-react", "^0.7.3"],
-                ["react", "^16.1.0"],
-                ["@types/react", "16.4.7"],
-                ["react-dom", "^16.2.0"],
+                ["msteams-ui-components-react", "^0.8.1"],
+                ["react", "^16.8.4"],
+                ["@types/react", "16.8.8"],
+                ["react-dom", "^16.8.4"],
                 ["file-loader", "1.1.11"],
-                ["typestyle", "1.5.1"]
+                ["typestyle", "2.0.1"]
             ], this.fs);
 
             // update client.ts
             Yotilities.insertTsExportDeclaration(
                 "src/app/scripts/client.ts",
-                `./${this.options.tabName}`,
+                `./${this.options.tabName}/${this.options.tabName}`,
                 `Automatically added for the ${this.options.tabName} tab`,
                 this.fs
             );
             Yotilities.insertTsExportDeclaration(
                 "src/app/scripts/client.ts",
-                `./${this.options.tabName}Config`,
+                `./${this.options.tabName}/${this.options.tabName}Config`,
                 undefined,
                 this.fs
             );
             Yotilities.insertTsExportDeclaration(
                 "src/app/scripts/client.ts",
-                `./${this.options.tabName}Remove`,
+                `./${this.options.tabName}/${this.options.tabName}Remove`,
                 undefined,
                 this.fs
             );